+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router } from "react-router-dom"
+import { PostsManager } from "./pages"
 import { Footer, Header } from "./widgets"
-import PostsManagerPage from "./pages/PostsManagerPage.tsx"
 
 const App = () => {
   return (
@@ -8,7 +8,7 @@ const App = () => {
       <div className="flex flex-col min-h-screen">
         <Header />
         <main className="flex-grow container mx-auto px-4 py-8">
-          <PostsManagerPage />
+          <PostsManager />
         </main>
         <Footer />
       </div>
