+++ src/feature/comment/model/index.ts
@@ -1,0 +1,3 @@
+export * from "./hook"
+export * from "./mutations"
+export * from "./store"
