+++ src/feature/comment/ui/index.ts
@@ -1,2 +1,3 @@
 export { CommentAddDialog } from "./CommentAddDialog"
 export { Comments } from "./Comments"
+export { CommentEditDialog } from "./CommentEditDialog"
