+++ src/feature/index.ts
@@ -1,1 +1,2 @@
 export * from "./comment"
+export * from "./post"
