+++ src/feature/post/index.ts
@@ -1,3 +1,3 @@
 export * from "./ui"
-export * from "./type"
 export * from "./model"
+export * from "./type"
