+++ src/feature/post/type.ts
@@ -1,5 +1,5 @@
 import { Post as originPost } from "../../entities"
-import { Author } from "../../shared/types"
+import { Author } from "../../shared"
 
 export interface Post extends originPost {
   author?: Author
