+++ src/pages/index.ts
@@ -1,0 +1,1 @@
+export {PostsManager} from './PostsManagerPage'
